+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for reading isyntax files using pyisyntax.
+
 ### Fixed
 
 - Reading correct image size in pyramid series in Bioformats reader.